memlat: Optimize perf event reads when possible
We can skip the locking and other overhead of perf_event_read_value()
when we know in advance that the perf event in question can be read from
the current CPU. This occurs when either the perf event permits reads
from CPUs other than the one its on, or when the CPU doing the reads is
the same CPU that owns the perf event.

Our PMU drivers only set two possible values for `readable_on_cpus`:
CPU_MASK_ALL or nothing. As such, we can simply check for CPU_MASK_ALL
beforehand in order to determine if the perf event allows non-local
reads.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[buzzcut: wip: adapt for 5.4]
Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>