pinctrl: msm: Remove explicit barriers from mmio ops where unneeded
For the vast majority of mmio operations in this driver, explicit memory
barriers aren't needed either because a data dependency between a read
and write already exists, or because of the presence of the spin locks
which execute a full memory barrier.

Removing all the unneeded explicit barriers considerably reduces
overhead for pinctrl operations, which in turn benefits things like i2c.

From:
https://github.com/kerneltoast/android_kernel_google_floral/commit/5f30e8d075d03ad490e81cb6a201cb3f94eb49de
https://github.com/Kaz205/renoir/commit/a66405a8b9960f1ba40a6469ab78fb94207534cd

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Kazuki Hashimoto <kazukih@tuta.io>