drm: Boost LLC bus upon running an atomic ioctl
In order to reduce jank, request a LLC bus boost as well whenever an
atomic ioctl is sent, which is what our display HAL uses to commit
frames to the display. The boost should be sufficient enough to render
the display's maximum FPS without any dropped frames when there is no
significant external source of load.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>