kernel: Boost LLC bus for a short amount of time when zygote forks
There is a gap from when a new app is forked to the time that the
devfreq governor managing the LLC bus can pick up the increased load and
ramp up appropriately. Since the polling interval for this governor is
typically 50 ms, that means it may take up to 50 ms before the governor
ramps up, which hurts cold launch app performance. Boosting the LLC bus
as well for 50 ms upon a fork from zygote covers that gap, speeding up
cold app launches.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>