Revert "memlat: Optimize perf event reads when possible"
This reverts commit a973e6f576f822135ec281d840654c2207b227c3.

There are a number of problems with this optimization attempt, such as how
a perf error code could be returned to one of the counter values, a CPU
could enter an SCM call for just one event and result in garbage
computations or a divide-by-zero when calculating the stall percentage,
and IRQs are disabled as a heavy-handed way of preventing CPU migration.

Toss this commit out so that it can be replaced by something better.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>