techpack: display: dsi: Affine dsi_ctrl IRQ to the big CPU cluster
OOS seems to forcefully affine this IRQ to the prime core. We get a
lot of interrupts from this IRQ (around half as many as the previously
affined msm_drm irq) which both keep the prime core not idle and
increase jitter.

I would much rather prefer it being affined to the litte cores. But
let's affine it to the big cluster for the time being instead.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>