techpack: display: dsi: Affine dsi_ctrl IRQ to the prime CPU cluster
OOS somehow manages to run dsi_ctrl on the same core as the DRM irq,
which we affine to the prime core [1].

Nothing short of reverting the DRM affinity commit fixes this issue.
Forcing dsi_ctrl to big CPUs also does not work [2][3].

Let's be explicit about the core affinity then - run it on prime.

[1]: https://github.com/buzzcut-s/android_kernel_oneplus_sm8350/commit/eb4b3bde119fd618b341f313362a1bdbb4277aad
[2]: https://github.com/buzzcut-s/android_kernel_oneplus_sm8350/commit/0062d4ec462fb62833fcdc6af47418fb1cffcc64
[3]: https://github.com/buzzcut-s/android_kernel_oneplus_sm8350/commit/6aa7de405a3456098aba60d253200786224e1d2c

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>